pyproject.toml
[project]
name = "real-time-logistics-strategy-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=45.0.0",
    "dagster>=1.12.20",
    "dagster-webserver>=1.12.20",
    "dbt-snowflake>=1.11.3",
    "faker>=40.11.0",
    "fastapi>=0.135.2",
    "numpy==2.4.2",
    "python-dotenv>=1.2.2",
    "snowflake-snowpark-python[pandas]>=1.47.0",
    "streamlit>=1.55.0",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.dagster]
module_name = "dagster_definitions"
code_location_name = "real-time-logistics-strategy-engine"

[tool.ruff]
line-length = 127
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.lint.isort]
combine-as-imports = true
